On logger

+++ Capstone/Classes/VendingMachine.cs
@@ -11,17 +11,17 @@ namespace Capstone.Classes
 
         //properties: balance, stock, 
         //methods: dispense item, list of items.  
-        public int Balance { get; set; }
+        public decimal Balance { get; set; }
 
         public List<Product> Products { get; set; } = new List<Product>();
         
 
-        public VendingMachine(int balance)
+        public VendingMachine(decimal balance)
         {
             Balance = balance;
         }
         //returns your final balance? 
-        public int FinalBalance()
+        public decimal FinalBalance()
         {
             return Balance;
         }
